# abroad-server/src/app/http/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"_36_Enums.TransactionStatus": {
				"type": "string",
				"enum": [
					"AWAITING_PAYMENT",
					"PROCESSING_PAYMENT",
					"PAYMENT_FAILED",
					"PAYMENT_EXPIRED",
					"PAYMENT_COMPLETED",
					"WRONG_AMOUNT"
				]
			},
			"DefaultSelection_Prisma._36_TransactionPayload_": {
				"properties": {
					"qrCode": {
						"type": "string"
					},
					"externalId": {
						"type": "string"
					},
					"taxId": {
						"type": "string"
					},
					"refundOnChainId": {
						"type": "string"
					},
					"onChainId": {
						"type": "string"
					},
					"quoteId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/_36_Enums.TransactionStatus"
					},
					"bankCode": {
						"type": "string"
					},
					"accountNumber": {
						"type": "string"
					},
					"partnerUserId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"qrCode",
					"externalId",
					"taxId",
					"refundOnChainId",
					"onChainId",
					"quoteId",
					"createdAt",
					"status",
					"bankCode",
					"accountNumber",
					"partnerUserId",
					"id"
				],
				"type": "object"
			},
			"Transaction": {
				"$ref": "#/components/schemas/DefaultSelection_Prisma._36_TransactionPayload_",
				"description": "Model Transaction"
			},
			"_36_Enums.TargetCurrency": {
				"type": "string",
				"enum": [
					"COP",
					"BRL"
				]
			},
			"TargetCurrency": {
				"$ref": "#/components/schemas/_36_Enums.TargetCurrency"
			},
			"PaginatedTransactionList": {
				"properties": {
					"page": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"transactions": {
						"items": {
							"allOf": [
								{
									"$ref": "#/components/schemas/Transaction"
								},
								{
									"properties": {
										"quote": {
											"properties": {
												"targetCurrency": {
													"$ref": "#/components/schemas/TargetCurrency"
												},
												"targetAmount": {
													"type": "number",
													"format": "double"
												},
												"sourceAmount": {
													"type": "number",
													"format": "double"
												},
												"paymentMethod": {
													"type": "string"
												},
												"network": {
													"type": "string"
												},
												"id": {
													"type": "string"
												},
												"cryptoCurrency": {
													"type": "string"
												}
											},
											"required": [
												"targetCurrency",
												"targetAmount",
												"sourceAmount",
												"paymentMethod",
												"network",
												"id",
												"cryptoCurrency"
											],
											"type": "object"
										}
									},
									"required": [
										"quote"
									],
									"type": "object"
								}
							]
						},
						"type": "array"
					}
				},
				"required": [
					"page",
					"pageSize",
					"total",
					"transactions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"QuoteResponse": {
				"properties": {
					"expiration_time": {
						"type": "number",
						"format": "double"
					},
					"quote_id": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"expiration_time",
					"quote_id",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.CryptoCurrency": {
				"type": "string",
				"enum": [
					"USDC"
				],
				"nullable": false
			},
			"CryptoCurrency": {
				"$ref": "#/components/schemas/_36_Enums.CryptoCurrency"
			},
			"_36_Enums.BlockchainNetwork": {
				"type": "string",
				"enum": [
					"STELLAR",
					"SOLANA"
				]
			},
			"BlockchainNetwork": {
				"$ref": "#/components/schemas/_36_Enums.BlockchainNetwork"
			},
			"_36_Enums.PaymentMethod": {
				"type": "string",
				"enum": [
					"NEQUI",
					"MOVII",
					"PIX",
					"BREB"
				]
			},
			"PaymentMethod": {
				"$ref": "#/components/schemas/_36_Enums.PaymentMethod"
			},
			"QuoteRequest": {
				"properties": {
					"target_currency": {
						"$ref": "#/components/schemas/TargetCurrency"
					},
					"payment_method": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"network": {
						"$ref": "#/components/schemas/BlockchainNetwork"
					},
					"crypto_currency": {
						"$ref": "#/components/schemas/CryptoCurrency"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"target_currency",
					"payment_method",
					"network",
					"crypto_currency",
					"amount"
				],
				"type": "object"
			},
			"ReverseQuoteRequest": {
				"properties": {
					"target_currency": {
						"$ref": "#/components/schemas/TargetCurrency"
					},
					"source_amount": {
						"type": "number",
						"format": "double"
					},
					"payment_method": {
						"$ref": "#/components/schemas/PaymentMethod"
					},
					"network": {
						"$ref": "#/components/schemas/BlockchainNetwork"
					},
					"crypto_currency": {
						"$ref": "#/components/schemas/CryptoCurrency"
					}
				},
				"required": [
					"target_currency",
					"source_amount",
					"payment_method",
					"network",
					"crypto_currency"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-Key"
			},
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "abroad",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/transactions/list": {
			"get": {
				"operationId": "ListPartnerTransactions",
				"responses": {
					"200": {
						"description": "Transactions retrieved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedTransactionList"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "List partner transactions (paginated)",
				"security": [
					{
						"ApiKeyAuth": []
					},
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "externalUserId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/quote": {
			"post": {
				"operationId": "GetQuote",
				"responses": {
					"200": {
						"description": "Quote response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuoteResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieves a quote to convert a given fiat amount into crypto.",
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-API-Key",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/QuoteRequest"
							}
						}
					}
				}
			}
		},
		"/quote/reverse": {
			"post": {
				"operationId": "GetReverseQuote",
				"responses": {
					"200": {
						"description": "Reverse quote response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QuoteResponse"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"reason": {
											"type": "string"
										}
									},
									"required": [
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieves a reverse quote: given the crypto amount the user sends,\nit returns the fiat amount (target amount) they would receive.",
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "X-API-Key",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReverseQuoteRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}